pyproject.toml
[project]
name = "walmart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "google-genai>=1.25.0",
    "numpy>=1.21.0",
    "opencv-python>=4.8.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "torch>=1.13.0",
    "torchvision>=0.14.0",
    "ultralytics>=8.0.0",
    "uvicorn>=0.35.0",
]
